+++ src/main/java/edu/epidata/empresa/controllers/EmpleadoController.java
@@ -25,7 +25,7 @@ public class EmpleadoController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Empleado> getEmpleado(@PathVariable(value="id")Integer id) {
+    public ResponseEntity<Empleado> getEmpleado(@PathVariable(value="id")Long id) {
         Optional<Empleado> empleado = empleadoService.findById(id);
         if (empleado.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -39,7 +39,7 @@ public class EmpleadoController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Empleado> updateEmpleado(@PathVariable(value="id")Integer id, @RequestBody Empleado empleado) {
+    public ResponseEntity<Empleado> updateEmpleado(@PathVariable(value="id")Long id, @RequestBody Empleado empleado) {
         Optional<Empleado> empleadoOptional = empleadoService.findById(id);
         if (empleadoOptional.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -48,7 +48,7 @@ public class EmpleadoController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Empleado> deleteEmpleado(@PathVariable(value="id")Integer id) {
+    public ResponseEntity<Empleado> deleteEmpleado(@PathVariable(value="id")Long id) {
         try {
             empleadoService.deleteById(id);
             return new ResponseEntity<>(HttpStatus.OK);
